Fixed tests to work as intended I think.
Spoke to Katharina and understood how the fits are supposed to work. Referenced ODMR logic which implements these fits. The main issue with the parameter generation seems to be that the x range is too large and the decay rate is generally too high which causes most of the data to be zeros and so the fit fails. I just reduced the x range significantly to fix this. Also need to re-implement fixtures or so to generate more parameter sets randomly. I removed them as part of my debugging steps because I couldn't get any fits to succeed.